refactor: reusable windows path fixing
